# Minor optimisations for `prepLookups`