Changed `prepLookups` to return a flat list instead of a list of lists. Also, added some newtypes like `PageSpan` and `PageNo`, and removed an indirection from the compact index `search` code.

I compared benchmarks before and after the change, using a fixed seed for the `prepLookups` micro-benchmark. I only benchmarked the default configuration with only positive lookups, because it exercises the most code (i.e., one bloom query and index search for each key-run combination).I also look at the allocations vs. iterations regression.

On 9020d6fe2870e7b137387d821fbc22aa7f71db1e:
```
benchmarking Bench.Database.LSMTree.Internal.Integration/prepLookups for a single run/default, ? actual pages/Bloomfilter query
time                 23.48 ms   (23.23 ms .. 23.67 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 24.24 ms   (24.02 ms .. 24.49 ms)
std dev              513.5 μs   (440.8 μs .. 614.9 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              1.284e8    (1.284e8 .. 1.284e8)
  y                  3557.754   (2426.069 .. 4857.572)

benchmarking Bench.Database.LSMTree.Internal.Integration/prepLookups for a single run/default, ? actual pages/Compact index search
time                 1.436 ms   (1.428 ms .. 1.446 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.489 ms   (1.478 ms .. 1.501 ms)
std dev              40.33 μs   (33.06 μs .. 47.93 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              335984.249 (335975.127 .. 335994.738)
  y                  3084.319   (2453.477 .. 3716.594)
variance introduced by outliers: 15% (moderately inflated)

benchmarking Bench.Database.LSMTree.Internal.Integration/prepLookups for a single run/default, ? actual pages/In-memory lookup
time                 26.08 ms   (25.90 ms .. 26.41 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 25.34 ms   (24.96 ms .. 25.64 ms)
std dev              756.0 μs   (550.2 μs .. 1.042 ms)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              1.332e8    (1.332e8 .. 1.332e8)
  y                  2140.211   (557.682 .. 3815.530)
```

On d6cffc092054630b761618bdae17049d5698805b:
```
benchmarking Bench.Database.LSMTree.Internal.Integration/prepLookups for a single run/default, ? actual pages/Bloomfilter query
time                 22.43 ms   (22.20 ms .. 22.69 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 22.67 ms   (22.58 ms .. 22.74 ms)
std dev              182.9 μs   (126.1 μs .. 253.0 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              1.284e8    (1.284e8 .. 1.284e8)
  y                  3510.737   (2411.617 .. 4691.646)

benchmarking Bench.Database.LSMTree.Internal.Integration/prepLookups for a single run/default, ? actual pages/Compact index search
time                 1.440 ms   (1.434 ms .. 1.445 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.485 ms   (1.474 ms .. 1.501 ms)
std dev              43.97 μs   (34.80 μs .. 54.69 μs)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              335984.612 (335975.994 .. 335993.636)
  y                  3062.105   (2468.657 .. 3687.059)
variance introduced by outliers: 17% (moderately inflated)

benchmarking Bench.Database.LSMTree.Internal.Integration/prepLookups for a single run/default, ? actual pages/In-memory lookup
time                 25.23 ms   (24.88 ms .. 25.60 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 24.11 ms   (23.65 ms .. 24.49 ms)
std dev              922.6 μs   (695.7 μs .. 1.221 ms)
allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
  iters              1.315e8    (1.315e8 .. 1.315e8)
  y                  2580.351   (1096.852 .. 4400.048)
```

Note that comparison will be rather imprecise, because the reported time allocation numbers are rounded. Time goes down a little it for "in-memory lookup", and allocations go down by roughly `0.017 * 10 ^ 8` . Though this may not look like much in the grand scheme of things, it shaves off roughly `1/3` of the allocations for `prepLookups` that can not be attributed to bloom filter queries or compact index searches.

These results also re-emphasise that bloom filter queries are quite costly