@jorisdral 2024-02-09
> Also nice to introduce the PageNo newtype, although the convenience of the Num instance takes away some of the benefits.


Removed the `Num` instance